pyproject.toml
[tool.poetry]
name = "customeraccounts"
version = "0.1.0"
description = ""
authors = ["James Groves <james.groves@shopify.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
black = "^24.8.0"
httpx = "^0.27.2"
pytest = "^8.3.2"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.6"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"
ruff = "^0.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
exclude = ["__init__.py"]
